requirements.txt
pyyaml==6.0
selenium==3.141.0
